Bugzilla #37815: Bugfix: The combination of content-width="scale-to-fit" and content-height="100%" did not work as expected due to a datatype conversion bug. Submitted by: Tom <tom.at.craddock.id.au>
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@354757 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/layoutmgr/inline/AbstractGraphicsLayoutManager.java
@@ -147,8 +147,8 @@ public abstract class AbstractGraphicsLayoutManager extends LeafNodeLayoutManage
                     cwidth = 0;
                     cheight = 0;
                 } else {
-                    double rat1 = cwidth / fobj.getIntrinsicWidth();
-                    double rat2 = cheight / fobj.getIntrinsicHeight();
+                    double rat1 = (double) cwidth / fobj.getIntrinsicWidth();
+                    double rat2 = (double) cheight / fobj.getIntrinsicHeight();
                     if (rat1 < rat2) {
                         // reduce cheight
                         cheight = (int)(rat1 * fobj.getIntrinsicHeight());
